WIP: Enable Java Applets: --enable-java=[yes|no]
I'd like to be able to use ShadowJS to support testing
JavaScript <=> Java Applet communication.
 
It seems QtWebKit might support Java Applets

This bug is marked as: RESOLVED FIXED
  Bug 33044 - Add Java Applet support to WebKit's Qt port
  https://bugs.webkit.org/show_bug.cgi?id=33044

This page reports: 
  http://dot.kde.org/2011/07/26/release-47-new-features-improved-stability-and-performance
  "... java applet support has been working since QtWebKit 2.1"

However a comment in qwebsettings.cpp indicates that 
currently Java applets are not supported.

./src/qt/src/3rdparty/webkit/Source/WebKit/qt/Api/qwebsettings.cpp

    \value JavaEnabled Enables or disables Java applets.
        Currently Java applets are not supported.

I ran the Jasmine tests on master and on the java-enabled branch
and both had the same single failure:

  #66 WebPage object: should handle file uploads
  FAIL

I did not see a current test pattern for testing an
argument like --load-plugins.

Using shadowjs with the  --enable-java argument appears
to operate without error, but rendering the page 
does not show the rendered Java applet.